// src/models/StarWarsCharacter.ts
export interface StarWarsCharacter {
    id: string;               // Identificador único (UUID)
    nombre: string;           // Nombre del personaje
    altura: string;           // Altura del personaje
    masa: string;             // Masa del personaje
    color_cabello: string;    // Color del cabello
    color_piel: string;       // Color de la piel
    color_ojos: string;       // Color de los ojos
    anio_nacimiento: string;  // Año de nacimiento
    genero: string;           // Género del personaje
  }  